Consider different architectures in the no-downgrade rule
Prevent cross-architecture downgrades (A:amd64=2 to A:i386=1)
by default, as they are somewhat surprising and can also
cause issues.

We should special case that for foreign/allowed, but doing
this more generically should not hurt either, as 'same' has
implicit same-version dependencies and 'no' only allows a
single installed version in a group anyway.