Only trust indices if they have a non-deprecated usable hash
This makes APT issue the usual "unauthenticated packages"
warning for packages coming from a Packages file that is
not listed with a usable non-deprecated hash in the Release
file.

It's a bit unfortunate that we have to parse the Release file
here, but there does not seem to be a better option right now.
Long term, we could introduce some kind of "<index>.untrusted"
file marking <index> as untrusted and avoid the parsing here.

We also need to adjust the tests a bit, as they force a hash
that is not present in a Release file and should thus fail
as well.