Run Pre-Install with fixed shell (default bash)
As explained in the previous commit: Despite a program catching the
signal, the shell might still exit with the signal when executing
sh -c "program", thus causing apt to think that the hook failed.

From testing, we know that dash is affected by this but mksh and
bash are not. As bash is essential, let's default to this, but
make it overridable at build time.

Bug-Debian: https://bugs.debian.org/832593