CMake: Explicitly list files instead of using GLOB_RECURSE
Using GLOB_RECURSE means that adding a new file or removing a
file is not detected automatically, causing issues in a distributed
development process:

Alice adds a new file, pushes it. Bob pulls the change and tries
to build, but gets missing symbols because the build process did
not rerun CMake and thus did not detect the new file.

Gbp-Dch: ignore