Allow increasing scores for manually installed packages
In practice we should really only consider automatic packages
an implementation detail: It does not matter for the user
which automatic packages are installed, as long as all manual
packages are installed.

The current scoring cannot really do this. This patch enables
us to boost the score of manually installed packages as a first
step. The default value is 0, though, so behavior won't change
for now - this is merely for playing.