Ensure connectivity for apt-daily using NM/networkd
Add systems that ensure that the network is up on systems
that use NetworkManager and/or systemd-networkd. These
are modelled after the normal wait-online services, but
do not contain RemainAfterExit=yes, so they are started
each time we start our service.

The time out is set to 1 hour which should work well
with 12 hour intervals.