Implement Happy Eyeballs for faster dual-stack connect
On dual-stack systems, IPv6 or IPv4 might not work correctly,
leading to timeouts. Our time out is 2 minutes, so it would take
2 * n minutes to try all IPv6 hosts before falling back to IPv4
if the IPv6 connections do not fail immediately (for example, because
the connect attemps are dropped rather than rejected).

This implements happy eyeballs: First, an IPv6 connect() attempt
is made. If that fails, and IPv4 attempt is made 300ms later, and
then we wait our normal timeout value (2 minutes) before trying
the next pair.

Closes: #668948