WIP: Happy eyeballs
This leaks file descriptors like hell, because adding Connections
to the vector fails if Connection has a destructor - it complains
about the copy constructor being deleted.

Adding a call to Close() in the MethodFd or FdFd destructor would
help, but apparently there is a uniqueness problem somewhere - the
fd gets closed while it should be used.