http: Fix handling of server connection closure
If the server closed the connection while we're reading data, and
we end up not having any data left to write; that is, for example,
we received 0 bytes, then we did not exit before, as we only returned
success if there was data to write.

This is wrong: Obviously, if we have reached our limit, we are done
anyway.

LP: #1801338