Store non-existing index targets as /dev/null symlinks
apt does a lot of stating on missing files, once uncompressed, and
then one time for each configured compressor. Store the files as
links to /dev/null instead and change existence check to use
RealFileExists() to exclude those symlinks from the cache (because
/dev/null is not a regular file).

This significantly reduces the number of stat() calls on a simple
apt-cache policy apt from 2271 calls (out of which 1827 were errors)
to 787 calls (with 29 errors). This might improve performance
considerably in places where syscalls are expensive.

Only do so for Packages files and Translations, so we avoid breakage
elsewhere.