Pass .deb files to store method for reverification
This makes sure that we reverify the deb files that we
received in case the http method did some funny things
like in CVE-2019-3462.

We also need to adjust the code for unsandboxed downloads
to unsandbox the store method if the file or copy method
is unsandboxed, as it will be accessing those files.