Do not send credentials after following cross-origin redirects
Whenever we do a redirect, note if we did a cross-origin redirect;
reset cross-origin redirect flag if we pop an alternative URI - while
it does not guarantee the alternative URI is being used, this seems
the easiest place, but it's fairly hacky.

When the HadCrossOriginRedirections flag is set, sent it to the method
in a new field Had-Cross-Origin-Redirections, and store it in the
FetchItem's HadCrossOriginRedirections.

Finally, If that flag is set, avoid calling MaybeAddAuth().

This completely breaks the expected behavior of netrc, which is to
work across redirects, so I'm not sure it's a sensible solution;
it's also quite messy and we might have a bunch of regressions pop
up because we are missing places.

Closes: #945911