pkgcache: Embed APT version and check it, set Major:Minor to 42:0
I forget to update the major:minor versions of the cache all the
time, so let's put an end to that and use the apt version directly.

This means that the cache needs rebuilding after apt is upgraded,
but this is a minor issue as it does not happen very often.