SECURITY UPDATE: Fix out of bounds read in .ar implementation (CVE-2020-3810)
If a member name consists solely of spaces or slashes, it causes
an out-of-bounds read. This might cause a denial of service or
a downgrade of service for services that use that code to inspect
deb files, such as Launchpad and dak, at the very least.

This rejects such members, and hence archives containing them;
dpkg, on the other hand, accepts empty file names. I don't believe
there is sufficient need for such file names, and we can't be sure
other code does not rely on filenames not being empty.

Fixes Debian/apt#111
LP: #1878177